Removes trailing slashes from URLs
URLs with trailing slashes are now redirected to their equivalent URL WITHOUT the trailing slash.
This trick leaves the security hole of allowing it to be possible for users to create usernames that conflict with site-specific paths, such as "users", "art", "journals", "meta", etc.
The hole is fixed by creating the "reserved_usernames" variable in config.py and filling it with the names of conflicting folders. If a user attempts to sign up with one of these reserved usernames, the signup fails and they are told that their username is already taken.